convert stat to TSX
