--- src/components/stat/index.js
@@ -1,3 +1,0 @@
-export {
-  EuiStat,
-} from './stat';
