+++ src/components/stat/index.ts
@@ -1,0 +1,1 @@
+export { EuiStat } from './stat';
